requirements.txt
boto3==1.34.162
python-dotenv==1.0.1